fix next tab flickering (#17)
When you say "next tab two times", vs code flickers between the two tabs
and only ends up moving one over. Originally we thought this might be
the UI delay being too short, but I saw this behavior at 250ms+, which
is too long. It looks like the problem is that we're focusing the old
tab because the active editor isn't updated frequently enough, so this
change makes focus check that first.